rename @types folder to types in order to fix nextjs build time errors
